fix sweat tests
